refactor(workspace): split useWorkspaceState god hook into focused slices
useWorkspaceState returned 45+ values from one 330-line hook. Extract three
cohesive slices — useWorkspaceUIState (panels/mobile view/chat width),
useWorkspaceNavigationState (selection/search/URL sync/derived values), and
useWorkspaceAIState (chat/token usage/AI edit) — and reduce
useWorkspaceState to a thin composition root. The public return shape is
unchanged, so consumers and tests are untouched.
